feat: always validate right operand of type parameter bounds
Previously, it was only done for the first lower/upper bound for each type parameter. The two issues are orthogonal to each other, however, and should thus both be shown.
