+++ packages/safe-ds-lang/src/language/validation/safe-ds-validator.ts
@@ -78,12 +78,15 @@ import {
     segmentResultMustBeAssignedExactlyOnce,
     segmentShouldBeUsed,
 } from './other/declarations/segments.js';
-import { typeParameterBoundLeftOperandMustBeOwnTypeParameter } from './other/declarations/typeParameterBounds.js';
+import {
+    typeParameterBoundLeftOperandMustBeOwnTypeParameter,
+    typeParameterBoundRightOperandMustBeNamedType,
+} from './other/declarations/typeParameterBounds.js';
 import {
     typeParameterBoundsMustBeCompatible,
     typeParameterMustBeUsedInCorrectPosition,
-    typeParameterMustHaveOneValidLowerAndUpperBound,
     typeParameterMustHaveSufficientContext,
+    typeParameterMustNotHaveMultipleBounds,
     typeParameterMustOnlyBeVariantOnClass,
 } from './other/declarations/typeParameters.js';
 import { callArgumentMustBeConstantIfParameterIsConstant, callMustNotBeRecursive } from './other/expressions/calls.js';
@@ -352,10 +355,13 @@ export const registerValidationChecks = function (services: SafeDsServices) {
             typeParameterBoundsMustBeCompatible(services),
             typeParameterMustBeUsedInCorrectPosition(services),
             typeParameterMustHaveSufficientContext,
-            typeParameterMustHaveOneValidLowerAndUpperBound(services),
+            typeParameterMustNotHaveMultipleBounds,
             typeParameterMustOnlyBeVariantOnClass,
         ],
-        SdsTypeParameterBound: [typeParameterBoundLeftOperandMustBeOwnTypeParameter],
+        SdsTypeParameterBound: [
+            typeParameterBoundLeftOperandMustBeOwnTypeParameter,
+            typeParameterBoundRightOperandMustBeNamedType(services),
+        ],
         SdsTypeParameterList: [
             typeParameterListMustNotHaveRequiredTypeParametersAfterOptionalTypeParameters,
             typeParameterListShouldNotBeEmpty(services),
